+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.5] - 2025-08-05
+
+### What's Changed
+- Add automatic capability handling for ICMP and UDP traceroute
+- No longer requires manual `sudo setcap` commands or sudo privileges
+- Add build script for setting capabilities during installation
+
 ## [0.1.4] - 2025-08-04
 
 ### What's Changed